src/main/java/com/aconsuegra/algorithms/ZeroMatrix.java
package com.aconsuegra.algorithms;

import java.util.HashSet;
import java.util.Set;

/**
 * Write an algorithm such that if an element in a matrix MxN is 0, it's entire row and column are set to 0
 */
public class ZeroMatrix {

    public int[][] setZero(int[][] matrix) {
        Set<Integer> zeroRows = new HashSet<>();
        Set<Integer> zeroColumns = new HashSet<>();
        for (int i = 0; i < matrix.length; i++) {
            for (int j = 0; j < matrix[i].length; j++) {
                if (matrix[i][j] == 0) {
                    zeroRows.add(i);
                    zeroColumns.add(j);
                }
            }
        }

        for (int i = 0; i < matrix.length; i++) {
            for (int j = 0; j < matrix[i].length; j++) {
                if (zeroRows.contains(i) || zeroColumns.contains(j)) {
                    matrix[i][j] = 0;
                }
            }
        }
        return matrix;
    }
}
